Fix Some bugs

+++ SysDev/SysDev/Controllers/SettingsController.cs
@@ -71,13 +71,22 @@ namespace SysDev.Controllers
 
             if (from.Equals("detail"))
             {
+                
                 var masterDetails = _context.MasterDetails.SingleOrDefault(p => p.Id == id);
+                var masterData = _context.MasterDatas.SingleOrDefault(md => md.Id == masterDetails.MasterDataId);
                 //var masterDetails = _context.MasterDetails.SingleOrDefault(p => p.MasterDataId == masterData.Id);
                 if (masterDetails == null)
                     return HttpNotFound();
 
+                var viewModel = new NewMasterDetailsViewModel
+                {
+                    MasterData = masterData,
+                    MasterDetail = masterDetails
+                };
+
+
                 ViewBag.ModalTitle = "Edit MasterDetails";
-                return View("CreateDetail",masterDetails);
+                return View("CreateDetail",viewModel);
             }
 
             return View();
@@ -91,10 +100,10 @@ namespace SysDev.Controllers
                 if (masterData != null)
                 {
                     masterData.Status = masterData.Status == "Active" ? "Inactive" : "Active";
-                    //ReportsController.AddAuditTrail("Update",
-                    //    "MasterData [ " + masterData.Name + "] was set to " + masterData.Status,
-                    //    User.Identity.GetUserId());
-                    //_context.SaveChanges();
+                    ReportsController.AddAuditTrail("Update",
+                        "MasterData [ " + masterData.Name + "] was set to " + masterData.Status,
+                        User.Identity.GetUserId());
+                    _context.SaveChanges();
                 }
             }
             else if (actionName.Equals("detail"))
@@ -103,30 +112,42 @@ namespace SysDev.Controllers
                 if (masterDetail != null)
                 {
                     masterDetail.Status = masterDetail.Status == "Active" ? "Inactive" : "Active";
-                    //ReportsController.AddAuditTrail("Update",
-                    //    "MasterDetail [ " + masterDetail.Name + "] was set to " + masterDetail.Status,
-                    //    User.Identity.GetUserId());
-                    //_context.SaveChanges();
+                    ReportsController.AddAuditTrail("Update",
+                        "MasterDetail [ " + masterDetail.Name + "] was set to " + masterDetail.Status,
+                        User.Identity.GetUserId());
+                    _context.SaveChanges();
                 }
             }
 
             return RedirectToAction("Index", "Settings");
         }
 
-        public ActionResult Delete(int id)
+        public ActionResult Delete(int id, string from)
         {
-            var masterData = _context.MasterDatas.SingleOrDefault(p => p.Id == id);
+            if (from.Equals("data"))
+            {
+                var masterData = _context.MasterDatas.SingleOrDefault(p => p.Id == id);
 
-            if (masterData == null)
-                return HttpNotFound();
+                if (masterData == null)
+                    return HttpNotFound();
 
-            _context.MasterDatas.Remove(masterData);
-            _context.SaveChanges();
+                _context.MasterDatas.Remove(masterData);
+                _context.SaveChanges();
 
-            //ReportsController.AddAuditTrail("Delete",
-            //    "MasterData [" + masterData.Name + "] was Deleted",
-            //    User.Identity.GetUserId());
+                ReportsController.AddAuditTrail("Delete",
+                    "MasterData [" + masterData.Name + "] was Deleted",
+                    User.Identity.GetUserId());
 
+            }
+            if(from.Equals("detail"))
+            {
+                var masterDetail = _context.MasterDetails.SingleOrDefault(p => p.Id == id);
+                if (masterDetail == null)
+                    return HttpNotFound();
+
+                _context.MasterDetails.Remove(masterDetail);
+                _context.SaveChanges();
+            }
             return RedirectToAction("Index", "Settings");
         }
 
@@ -158,9 +179,9 @@ namespace SysDev.Controllers
                 model.Status = "Active";
                 _context.MasterDatas.Add(model);
                 _context.SaveChanges();
-                //ReportsController.AddAuditTrail("Add MasterData",
-                //    model.Name + " has been added.",
-                //    User.Identity.GetUserId());
+                ReportsController.AddAuditTrail("Add MasterData",
+                    model.Name + " has been added.",
+                    User.Identity.GetUserId());
             }
             else
             {
@@ -171,9 +192,9 @@ namespace SysDev.Controllers
 
                     mData.Description = model.Description;
                     mData.DateTimeUpdated = DateTime.Now;
-                    //ReportsController.AddAuditTrail("Update",
-                    //    "MasterData [" + model.Name + "] has been added.",
-                    //    User.Identity.GetUserId());
+                    ReportsController.AddAuditTrail("Update",
+                        "MasterData [" + model.Name + "] has been added.",
+                        User.Identity.GetUserId());
                 }
 
             }
@@ -184,9 +205,9 @@ namespace SysDev.Controllers
 
         public ActionResult SaveMasterDetail(NewMasterDetailsViewModel model)
         {
-            var masterData = _context.MasterDatas.SingleOrDefault(m => m.Id == model.MasterData.Id);
-            if (model.MasterDetail.Id == 0 && masterData !=null)
+            if (model.MasterDetail.Id == 0)
             {
+                var masterData = _context.MasterDatas.SingleOrDefault(m => m.Id == model.MasterData.Id);
                 var masterDetail = new MasterDetail
                 {
                     MasterData = masterData,
@@ -198,6 +219,7 @@ namespace SysDev.Controllers
                     Status = "Active",
                     Value = model.MasterDetail.Name
                 };
+                _context.MasterDetails.Add(masterDetail);
             }
             else
             {
